work in progress for RIP across interfaces to route table

+++ examples/net_rip.py
@@ -4,8 +4,8 @@ import sys
 sys.path.append("..")
 from anhost import anhost# for all the active node things
 
-dst = "172.31.13.99"
-src = anhost.get_ip_address("eth0")
+src = anhost.get_ip_address("eth1")
+dst = src
 
 # AN port
 port = 50000
@@ -56,29 +56,32 @@ signal.signal(signal.SIGHUP, signal_term_handler)
 
 logger.debug("Checking for File")
 ## if network server not running, start it
-if not os.path.exists(route_fi):
-  logger.debug("FILE DOES NOT EXIST: %s" % route_fi)
-  rip_thread = threading.Thread(target=rip.rip_server, args=(open(fi).read(),\
-                                serv_port,net_port,route_fi,))
-  try:
-    fi_o = open(route_fi,'w')
-    fi_o.close()
-    logger.debug("%s Thread: RIP Server" % fi)
-    logger.debug("PID: %s" % os.getpid())
-    #rip_thread.daemon = True
-    rip_thread.start()
-    rip_thread.join()
-    logger.debug("RIP Server started")
-  except KeyboardInterrupt:
-    logger.error("Key recieved! Killing threads")
-    rip_thread.kill()
-    exit(10)
-  except Exception,e:
-    logger.error("error in %s: %s" % (fi,str(e)))
-    logger.error("deleting server file: %s" % route_fi)
-    if os.path.exists(route_fi):
-      os.remove(route_fi)
-      raise Exception(e)
+iface_list = anhost.non_default_routes()
+for iface in iface_list:
+  route_fi+=str(iface["Iface"])
+  if not os.path.exists(route_fi) and iface != "eth0":
+    logger.debug("FILE DOES NOT EXIST: %s" % route_fi)
+    rip_thread = threading.Thread(target=rip.rip_server, args=(open(fi).read(),\
+                                  serv_port,net_port,route_fi,iface,))
+    try:
+      fi_o = open(route_fi,'w')
+      fi_o.close()
+      logger.debug("%s Thread: RIP Server" % fi)
+      logger.debug("PID: %s" % os.getpid())
+      #rip_thread.daemon = True
+      rip_thread.start()
+      rip_thread.join()
+      logger.debug("RIP Server started")
+    except KeyboardInterrupt:
+      logger.error("Key recieved! Killing threads")
+      rip_thread.kill()
+      exit(10)
+    except Exception,e:
+      logger.error("error in %s: %s" % (fi,str(e)))
+      logger.error("deleting server file: %s" % route_fi)
+      if os.path.exists(route_fi):
+        os.remove(route_fi)
+        raise Exception(e)
     
 else:
   logger.debug("ROUTE FILE EXISTS- EXIT")
@@ -87,6 +90,5 @@ exit(0)
 """
 
 # we should run it locally first to populate, send to self first, then forward.
-local = anhost.get_ip_address("eth0")
 local_code = anhost.set_src_dst(AN_code, src,dst)
-sock.sendto(local_code, (local,port))
+sock.sendto(local_code, (src,port))
